split AEPPushNotificationBuilder methods to extension functions
