+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/builders/AutoCarouselNotificationBuilder.kt
@@ -12,6 +12,7 @@
 package com.adobe.marketing.mobile.services.ui.notification.builders
 
 import android.app.Activity
+import android.app.NotificationManager
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.graphics.Bitmap
@@ -24,6 +25,8 @@ import com.adobe.marketing.mobile.services.caching.CacheService
 import com.adobe.marketing.mobile.services.ui.notification.NotificationConstructionFailedException
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateConstants
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateImageUtil
+import com.adobe.marketing.mobile.services.ui.notification.extensions.createNotificationChannelIfRequired
+import com.adobe.marketing.mobile.services.ui.notification.extensions.setRemoteViewClickAction
 import com.adobe.marketing.mobile.services.ui.notification.templates.AutoCarouselPushTemplate
 import com.adobe.marketing.mobile.services.ui.notification.templates.CarouselPushTemplate
 
@@ -56,8 +59,10 @@ internal object AutoCarouselNotificationBuilder {
         val smallLayout = RemoteViews(packageName, R.layout.push_template_collapsed)
         val expandedLayout = RemoteViews(packageName, R.layout.push_template_auto_carousel)
 
+        val notificationManager =
+            context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
         // Create the notification channel if needed
-        val channelIdToUse = AEPPushNotificationBuilder.createChannelIfRequired(
+        val channelIdToUse = notificationManager.createNotificationChannelIfRequired(
             context,
             pushTemplate.channelId,
             pushTemplate.sound,
@@ -157,10 +162,9 @@ internal object AutoCarouselNotificationBuilder {
             // assign a click action pending intent for each carousel item if we have a tracker activity
             trackerActivityClass?.let {
                 val interactionUri = item.interactionUri ?: pushTemplate.actionUri
-                AEPPushNotificationBuilder.setRemoteViewClickAction(
+                carouselItem.setRemoteViewClickAction(
                     context,
                     trackerActivityClass,
-                    carouselItem,
                     R.id.carousel_item_image_view,
                     interactionUri,
                     pushTemplate.tag,
