+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/builders/BasicNotificationBuilder.kt
@@ -13,11 +13,8 @@ package com.adobe.marketing.mobile.services.ui.notification.builders
 
 import android.app.Activity
 import android.app.NotificationManager
-import android.app.PendingIntent
 import android.content.BroadcastReceiver
 import android.content.Context
-import android.content.Intent
-import android.os.Build
 import android.view.View
 import android.widget.RemoteViews
 import androidx.core.app.NotificationCompat
@@ -25,13 +22,12 @@ import com.adobe.marketing.mobile.core.R
 import com.adobe.marketing.mobile.services.Log
 import com.adobe.marketing.mobile.services.ServiceProvider
 import com.adobe.marketing.mobile.services.ui.notification.NotificationConstructionFailedException
+import com.adobe.marketing.mobile.services.ui.notification.PendingIntentUtils
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateConstants
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateImageUtil
-import com.adobe.marketing.mobile.services.ui.notification.templates.AEPPushTemplate
+import com.adobe.marketing.mobile.services.ui.notification.extensions.addActionButtons
+import com.adobe.marketing.mobile.services.ui.notification.extensions.createNotificationChannelIfRequired
 import com.adobe.marketing.mobile.services.ui.notification.templates.BasicPushTemplate
-import org.json.JSONArray
-import org.json.JSONException
-import org.json.JSONObject
 
 /**
  * Object responsible for constructing a [NotificationCompat.Builder] object containing a basic push template notification.
@@ -58,22 +54,14 @@ internal object BasicNotificationBuilder {
         val smallLayout = RemoteViews(packageName, R.layout.push_template_collapsed)
         val expandedLayout = RemoteViews(packageName, R.layout.push_template_expanded)
 
-        // create a silent notification channel if needed
-        if (pushTemplate.isFromIntent == true && Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val notificationManager =
-                context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            AEPPushNotificationBuilder.setupSilentNotificationChannel(
-                notificationManager,
-                pushTemplate.getNotificationImportance()
-            )
-        }
-
-        // create the notification channel if needed
-        val channelIdToUse = AEPPushNotificationBuilder.createChannelIfRequired(
+        val notificationManager =
+            context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        val channelIdToUse: String = notificationManager.createNotificationChannelIfRequired(
             context,
             pushTemplate.channelId,
             pushTemplate.sound,
-            pushTemplate.getNotificationImportance()
+            pushTemplate.getNotificationImportance(),
+            pushTemplate.isFromIntent
         )
 
         // create the notification builder with the common settings applied
@@ -110,10 +98,9 @@ internal object BasicNotificationBuilder {
         )
 
         // add any action buttons defined for the notification
-        addActionButtons(
+        notificationBuilder.addActionButtons(
             context,
             trackerActivityClass,
-            notificationBuilder,
             pushTemplate.actionButtonsString,
             pushTemplate.tag,
             pushTemplate.isNotificationSticky ?: false
@@ -124,246 +111,17 @@ internal object BasicNotificationBuilder {
             if (pushTemplate.remindLaterEpochTimestamp != null ||
                 pushTemplate.remindLaterDelaySeconds != null
             ) {
-                val remindIntent =
-                    createRemindPendingIntent(
-                        context,
-                        broadcastReceiverClass,
-                        channelIdToUse,
-                        pushTemplate
-                    )
+                val remindIntent = PendingIntentUtils.createRemindPendingIntent(
+                    context,
+                    broadcastReceiverClass,
+                    channelIdToUse,
+                    pushTemplate
+                )
                 notificationBuilder.addAction(0, remindLaterText, remindIntent)
             }
         }
 
         return notificationBuilder
-    }
-
-    /**
-     * Adds action buttons for the notification.
-     *
-     * @param context the application [Context]
-     * @param trackerActivityClass the [Activity] class to use as the tracker activity
-     * @param builder the [NotificationCompat.Builder] to attach the action buttons
-     * @param actionButtonsString `String` a JSON string containing action buttons to attach
-     * to the notification
-     * notification
-     * @param tag `String` containing the tag to use when scheduling the notification
-     * @param stickyNotification [Boolean]  if false, remove the notification after the action
-     * button is pressed
-     */
-    internal fun addActionButtons(
-        context: Context,
-        trackerActivityClass: Class<out Activity>?,
-        builder: NotificationCompat.Builder,
-        actionButtonsString: String?,
-        tag: String?,
-        stickyNotification: Boolean
-    ) {
-        val actionButtons: List<AEPPushTemplate.ActionButton>? =
-            getActionButtonsFromString(actionButtonsString)
-        if (actionButtons.isNullOrEmpty()) {
-            return
-        }
-        for (eachButton in actionButtons) {
-            val pendingIntent: PendingIntent? =
-                if (eachButton.type === PushTemplateConstants.ActionType.DEEPLINK ||
-                    eachButton.type === PushTemplateConstants.ActionType.WEBURL
-                ) {
-                    AEPPushNotificationBuilder.createPendingIntent(
-                        context,
-                        trackerActivityClass,
-                        eachButton.link,
-                        eachButton.label,
-                        tag,
-                        stickyNotification
-                    )
-                } else {
-                    AEPPushNotificationBuilder.createPendingIntent(
-                        context,
-                        trackerActivityClass,
-                        null,
-                        eachButton.label,
-                        tag,
-                        stickyNotification
-                    )
-                }
-            builder.addAction(0, eachButton.label, pendingIntent)
-        }
-    }
-
-    private fun getActionButtonsFromString(actionButtons: String?): List<AEPPushTemplate.ActionButton>? {
-        if (actionButtons == null) {
-            Log.debug(
-                PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Exception in converting actionButtons json string to json object, Error :" +
-                    " actionButtons is null"
-            )
-            return null
-        }
-        val actionButtonList = mutableListOf<AEPPushTemplate.ActionButton>()
-        try {
-            val jsonArray = JSONArray(actionButtons)
-            for (i in 0 until jsonArray.length()) {
-                val jsonObject = jsonArray.getJSONObject(i)
-                val button = getActionButton(jsonObject) ?: continue
-                actionButtonList.add(button)
-            }
-        } catch (e: JSONException) {
-            Log.warning(
-                PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Exception in converting actionButtons json string to json object, Error : ${e.localizedMessage}"
-            )
-            return null
-        }
-        return actionButtonList
-    }
-
-    private fun getActionButton(jsonObject: JSONObject): AEPPushTemplate.ActionButton? {
-        return try {
-            val label = jsonObject.getString(PushTemplateConstants.ActionButtons.LABEL)
-            if (label.isEmpty()) {
-                Log.debug(PushTemplateConstants.LOG_TAG, SELF_TAG, "Label is empty")
-                return null
-            }
-            var uri: String? = null
-            val type = jsonObject.getString(PushTemplateConstants.ActionButtons.TYPE)
-            if (type == PushTemplateConstants.ActionType.WEBURL.name || type == PushTemplateConstants.ActionType.DEEPLINK.name) {
-                uri = jsonObject.optString(PushTemplateConstants.ActionButtons.URI)
-            }
-            Log.trace(
-                PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Creating an ActionButton with label ($label), uri ($uri), and type ($type)."
-            )
-            AEPPushTemplate.ActionButton(label, uri, type)
-        } catch (e: JSONException) {
-            Log.warning(
-                PushTemplateConstants.LOG_TAG,
-                SELF_TAG,
-                "Exception in converting actionButtons json string to json object, Error : ${e.localizedMessage}."
-            )
-            null
-        }
-    }
-
-    private fun createRemindPendingIntent(
-        context: Context,
-        broadcastReceiverClass: Class<out BroadcastReceiver>?,
-        channelId: String,
-        pushTemplate: BasicPushTemplate
-    ): PendingIntent? {
-        if (broadcastReceiverClass == null) {
-            return null
-        }
-        Log.trace(
-            PushTemplateConstants.LOG_TAG,
-            SELF_TAG,
-            "Creating a remind later pending intent from a push template object."
-        )
-
-        val remindIntent = Intent(PushTemplateConstants.IntentActions.REMIND_LATER_CLICKED)
-        broadcastReceiverClass.let {
-            remindIntent.setClass(context.applicationContext, broadcastReceiverClass)
-        }
-
-        remindIntent.flags = Intent.FLAG_ACTIVITY_SINGLE_TOP
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.TEMPLATE_TYPE, pushTemplate.templateType?.value
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.IMAGE_URI, pushTemplate.imageUrl
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.ACTION_URI, pushTemplate.actionUri
-        )
-        remindIntent.putExtra(PushTemplateConstants.IntentKeys.CHANNEL_ID, channelId)
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.CUSTOM_SOUND, pushTemplate.sound
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.TITLE_TEXT,
-            pushTemplate.title
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.BODY_TEXT,
-            pushTemplate.body
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT,
-            pushTemplate.expandedBodyText
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.NOTIFICATION_BACKGROUND_COLOR,
-            pushTemplate.notificationBackgroundColor
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.TITLE_TEXT_COLOR,
-            pushTemplate.titleTextColor
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.EXPANDED_BODY_TEXT_COLOR,
-            pushTemplate.expandedBodyTextColor
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.SMALL_ICON, pushTemplate.smallIcon
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.SMALL_ICON_COLOR,
-            pushTemplate.smallIconColor
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.LARGE_ICON, pushTemplate.largeIcon
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.VISIBILITY,
-            pushTemplate.getNotificationVisibility()
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.IMPORTANCE,
-            pushTemplate.getNotificationImportance()
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.BADGE_COUNT, pushTemplate.badgeCount
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.REMIND_EPOCH_TS, pushTemplate.remindLaterEpochTimestamp
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.REMIND_DELAY_SECONDS,
-            pushTemplate.remindLaterDelaySeconds
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.REMIND_LABEL, pushTemplate.remindLaterText
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.ACTION_BUTTONS_STRING,
-            pushTemplate.actionButtonsString
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.STICKY, pushTemplate.isNotificationSticky
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.TAG, pushTemplate.tag
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.TICKER, pushTemplate.ticker
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.PAYLOAD_VERSION, pushTemplate.payloadVersion
-        )
-        remindIntent.putExtra(
-            PushTemplateConstants.IntentKeys.PRIORITY,
-            pushTemplate.notificationPriority
-        )
-
-        return PendingIntent.getBroadcast(
-            context,
-            0,
-            remindIntent,
-            PendingIntent.FLAG_MUTABLE or PendingIntent.FLAG_UPDATE_CURRENT
-        )
     }
 
     @Throws(NotificationConstructionFailedException::class)
