+++ code/core/src/phone/java/com/adobe/marketing/mobile/services/ui/notification/builders/LegacyNotificationBuilder.kt
@@ -12,11 +12,19 @@
 package com.adobe.marketing.mobile.services.ui.notification.builders
 
 import android.app.Activity
+import android.app.NotificationManager
 import android.content.Context
 import android.os.Build
 import androidx.core.app.NotificationCompat
 import com.adobe.marketing.mobile.services.Log
 import com.adobe.marketing.mobile.services.ui.notification.PushTemplateConstants
+import com.adobe.marketing.mobile.services.ui.notification.extensions.addActionButtons
+import com.adobe.marketing.mobile.services.ui.notification.extensions.createNotificationChannelIfRequired
+import com.adobe.marketing.mobile.services.ui.notification.extensions.setLargeIcon
+import com.adobe.marketing.mobile.services.ui.notification.extensions.setNotificationClickAction
+import com.adobe.marketing.mobile.services.ui.notification.extensions.setNotificationDeleteAction
+import com.adobe.marketing.mobile.services.ui.notification.extensions.setSmallIcon
+import com.adobe.marketing.mobile.services.ui.notification.extensions.setSound
 import com.adobe.marketing.mobile.services.ui.notification.templates.BasicPushTemplate
 
 /**
@@ -37,7 +45,9 @@ internal object LegacyNotificationBuilder {
         )
 
         // create the notification channel
-        val channelId = AEPPushNotificationBuilder.createChannelIfRequired(
+        val notificationManager =
+            context.getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        val channelId = notificationManager.createNotificationChannelIfRequired(
             context,
             pushTemplate.channelId,
             pushTemplate.sound,
@@ -45,65 +55,41 @@ internal object LegacyNotificationBuilder {
         )
 
         // Create the notification builder object and set the ticker, title, body, and badge count
-        val builder = NotificationCompat.Builder(
-            context, channelId
-        )
+        val builder = NotificationCompat.Builder(context, channelId)
             .setTicker(pushTemplate.ticker)
             .setContentTitle(pushTemplate.title)
             .setContentText(pushTemplate.body)
             .setNumber(pushTemplate.badgeCount)
-
-        // set a large icon if one is present
-        AEPPushNotificationBuilder.setLargeIcon(
-            builder,
-            pushTemplate.imageUrl,
-            pushTemplate.title,
-            pushTemplate.expandedBodyText
-        )
-
-        // small Icon must be present, otherwise the notification will not be displayed.
-        AEPPushNotificationBuilder.setSmallIcon(
-            context,
-            builder,
-            pushTemplate.smallIcon,
-            pushTemplate.smallIconColor
-        )
-
-        // set notification visibility
-        AEPPushNotificationBuilder.setVisibility(
-            builder, pushTemplate.getNotificationVisibility()
-        )
-
-        // add any action buttons defined for the notification
-        BasicNotificationBuilder.addActionButtons(
-            context,
-            trackerActivityClass,
-            builder,
-            pushTemplate.actionButtonsString,
-            pushTemplate.tag,
-            pushTemplate.isNotificationSticky ?: false
-        )
-
-        // set custom sound, note this applies to API 25 and lower only as API 26 and up set the
-        // sound on the notification channel
-        AEPPushNotificationBuilder.setSound(context, builder, pushTemplate.sound)
-
-        // assign a click action pending intent to the notification
-        AEPPushNotificationBuilder.setNotificationClickAction(
-            context,
-            trackerActivityClass,
-            builder,
-            pushTemplate.actionUri,
-            pushTemplate.tag,
-            pushTemplate.isNotificationSticky ?: false
-        )
-
-        // set notification delete action
-        AEPPushNotificationBuilder.setNotificationDeleteAction(
-            context,
-            trackerActivityClass,
-            builder
-        )
+            // set a large icon if one is present
+            .setLargeIcon(pushTemplate.imageUrl, pushTemplate.title, pushTemplate.expandedBodyText)
+            // small Icon must be present, otherwise the notification will not be displayed.
+            .setSmallIcon(context, pushTemplate.smallIcon, pushTemplate.smallIconColor)
+            // set notification visibility
+            .setVisibility(pushTemplate.getNotificationVisibility())
+            // add any action buttons defined for the notification
+            .addActionButtons(
+                context,
+                trackerActivityClass,
+                pushTemplate.actionButtonsString,
+                pushTemplate.tag,
+                pushTemplate.isNotificationSticky ?: false
+            )
+            // set custom sound, note this applies to API 25 and lower only as API 26 and up set the
+            // sound on the notification channel
+            .setSound(context, pushTemplate.sound)
+            // assign a click action pending intent to the notification
+            .setNotificationClickAction(
+                context,
+                trackerActivityClass,
+                pushTemplate.actionUri,
+                pushTemplate.tag,
+                pushTemplate.isNotificationSticky ?: false
+            )
+            // set notification delete action
+            .setNotificationDeleteAction(
+                context,
+                trackerActivityClass
+            )
 
         // if API level is below 26 (prior to notification channels) then notification priority is
         // set on the notification builder
